agent/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-dotenv==1.1.0
python-multipart==0.0.20
pydantic==2.11.3
openai==1.73.0
openai-agents==0.0.9
httpx==0.28.1
aiofiles==24.1.0 